Start work on an enclosed SpawnMonitor module
This module will handle as much of the incursion spawn detection internally as possible, only relying on external services such as REDIS, EveSwagger, and DiscordWebhook when needed.

To start this project off a new folder has been created, with a new extension added to the service collection. The old Discord Controller has been deleted, and a new one created which makes use of the new SpawnWebhook model which includes additional settings.

Next up is to move jobs to the module namespace.